Time Ability Component added
Gives the player the ability to slow time by pressing a button (currently Right Mouse Button, but this can be changed in the IMC_RailCharacter asset if needed). This works by using at the desired changing the global time dilation, and then we decrease/increase our meter on a timer every [MeterUseRate] seconds. Timer only runs while ability is in use/recharging. All parameters customizable in component blueprint attached to the player.